Fixed broken AggressorAI
Former-commit-id: 6ba2768a409b7b09b319bad8ec4e7192d6765bb6

+++ src/game/ObjectMgr.cpp
@@ -1472,7 +1472,7 @@ void ObjectMgr::LoadCreatureRespawnTimes()
         Field *fields = result->Fetch();
 
         uint32 loguid       = fields[0].GetUInt32();
-        uint64 respawn_time = fields[1].GetUInt64();
+        uint64 respawn_time = fields[1].GetUInt32();
         uint32 instance     = fields[2].GetUInt32();
 
         mCreatureRespawnTimes[MAKE_PAIR64(loguid,instance)] = time_t(respawn_time);
